ADD support for default or namespace-based Hive databases creation
