+++ CHANGES_NEXT_RELEASE
@@ -4,6 +4,7 @@
 - [HARDENING] Remove the dual batch accumulation in OrionSink (#664)
 - [HARDENING] Add "requests" package as a requirement for cygnus_translator_pre0.10.0_to_0.10.0_ckan.py (#667)
 - [HARDENING] Allow SSL connections to CKAN servers in cygnus_translator_pre0.10.0_to_0.10.0_ckan.py (#666)
-- [HARDENING] Add the maximum accepted batch size in DynamoSink (#672)
+- [HARDENING] Add the maximum accepted batch size in OrionDynamoDBSink (#672)
 - [HARDENING] Add interface and implementation to MongoBackend (#654)
 - [HARDENING] Add grouping_rules.conf & flume-env.sh configuration files in the documentation (#657)
+- [FEATURE] Enable per-user or default Hive database creation in OrionHDFSSink(#516)
